feat(oxidize-c): implement keyframe sampling + video encoder GEMV projection (126K lines, 2735 tests)
Frame sampler (replaces uniform fallback):
- OC_FS_STRATEGY_KEYFRAME: real keyframe sampling that always includes
  first and last frame, distributes remaining frames evenly across the
  video for maximum temporal coverage

Video encoder (replaces copy/pad stub):
- OcVideoEncoder extended with proj_weight [llm_hidden, vision_hidden]
  and proj_bias [llm_hidden] fields
- oc_video_encoder_encode: when proj_weight available, performs real
  GEMV projection (out = W @ frame + bias); falls back to copy/pad
  when no weights loaded
- Proper cleanup in oc_video_encoder_free for new weight fields

+++ oxidize-c/tests/test_inf_forward.c
@@ -673,7 +673,7 @@ Test(inf_fwd, forward_batch_basic)
 
     /* Create 2 sequences with their own SeqKv buffers. */
     size_t kv_layer_count = m.config.layer_count;
-    size_t kv_len = m.config.num_key_value_heads * oc_inference_config_kv_head_dim(&m.config);
+    size_t kv_len = (size_t)m.config.num_key_value_heads * oc_inference_config_kv_head_dim(&m.config);
     size_t cap = 32;
 
     OcSeqKv kv0, kv1;
